Add the default header provided by MUI.
In this commit, I created a folder that will store my components. I will most likely make changes to how components/pages are organized in the future. Furthermore, I added the default header that MUI provides. I will make changes to this in the future so that it works for my project. Additionally, I installed MUI to use the header on the website.